% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findr.R
\name{findr}
\alias{findr}
\title{Characterize a gene using public genomic databases}
\usage{
findr(gene, disease = NULL, site = NULL)
}
\arguments{
\item{gene}{Character. Gene symbol (e.g. "TP53")}

\item{disease}{Character. Disease context (e.g. "alzheimer")}

\item{site}{Character. Cancer site (e.g. "breast", "lung"). Can be a vector for multi-site comparison.}
}
\value{
Invisibly returns a data frame of results
}
\description{
Characterize a gene using public genomic databases
}
\examples{
\donttest{
findr("TP53", site = "breast")
}
}
