% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findr.R
\name{findr_multi}
\alias{findr_multi}
\title{Characterize multiple genes using public genomic databases}
\usage{
findr_multi(genes, disease = NULL, site = NULL, output = "print")
}
\arguments{
\item{genes}{Character vector. Gene symbols (e.g. c("TP53", "BRCA1"))}

\item{disease}{Character. Disease context}

\item{site}{Character. Cancer site}

\item{output}{Character. "print" or "table"}
}
\value{
Invisibly returns a data frame of results
}
\description{
Characterize multiple genes using public genomic databases
}
\examples{
\donttest{
findr_multi(c("TP53", "BRCA1"), site = "breast")
}
}
